Maia loading works
Test script includes a simplified function for loading Maia networks.
The main change in the Maia is the removal of an argument for creating an engine which caused it to crash.
In addition an engine running cuda is included in the lc0-v0.30.0-windows-gpu-nvidia-cuda folder, if your machine does not have a gpu with cuda this should be replaced by a different one found on:
https://lczero.org/play/download/